requirements.txt
pytest==4.6.2
facebook-sdk==3.1.0
requests==2.22.0